Se posicionan los personajes de manera independiente a PesonajeDef, sino que a travez de personajes pertenecientes a cada jugador respectivo

+++ src/modelo/Vector2f.cpp
@@ -32,7 +32,7 @@ Vector2f& Vector2f::operator*=(float escalar) {
     return *this;
 }
 
-void Vector2f::setCoordenada(int x, int y) {
+void Vector2f::setCoordenada(double x, double y) {
     this->x = x;
     this->y = y;
 }
